PORTAL-56 move section files into sections folder
