+++ src/components/OnBoardingForm.jsx
@@ -1,13 +1,13 @@
 import PropTypes from "prop-types";
-import { STATUS, SECTION } from "../constants/onBoarding";
-import CarInformation from "./CarInformation";
-import ContactInformation from "./ContactInformation";
-import CurrentAddress from "./CurrentAddress";
-import DriverLicense from "./DriverLicense";
-import EmergencyContacts from "./EmergencyContacts";
-import ImmigrationInformation from "./ImmigrationInformation";
-import PersonalDetails from "./PersonalDetails";
-import Reference from "./Reference";
+import { SECTION } from "../constants/onBoarding";
+import CarInformation from "./sections/CarInformation";
+import ContactInformation from "./sections/ContactInformation";
+import CurrentAddress from "./sections/CurrentAddress";
+import DriverLicense from "./sections/DriverLicense";
+import EmergencyContacts from "./sections/EmergencyContacts";
+import ImmigrationInformation from "./sections/ImmigrationInformation";
+import PersonalDetails from "./sections/PersonalDetails";
+import Reference from "./sections/Reference";
 function OnBoardingForm({ disabled, section }) {
   switch (section) {
     case SECTION.CAR_INFORMATION:
