+++ src/components/sections/Reference.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import EmergencyContact from "./EmergencyContact";
+import EmergencyContact from "../EmergencyContact";
 import { useDispatch, useSelector } from "react-redux";
-import { selectReference } from "../store/onBoardingSlice/onBoarding.selectors";
-import { updateReference } from "../store/onBoardingSlice/onBoarding.slice";
+import { selectReference } from "../../store/onBoardingSlice/onBoarding.selectors";
+import { updateReference } from "../../store/onBoardingSlice/onBoarding.slice";
 function Reference({ disabled = false }) {
   const reference = useSelector(selectReference);
   const dispatch = useDispatch();
